Consolidate extensions definition into a single file
